requirements.txt
asgiref
Django
djangorestframework
python-decouple
sqlparse
typing_extensions
gunicorn
whitenoise
